fs: crypto: Correct erroneous -Wmisleading-indentation resolution
In commit a8d76e4 ("treewide: Fix remaining misleading-indentation warnings"), we attempted to fix the following compilation warning:

../fs/crypto/policy.c:636:4: warning: misleading indentation; statement is not part of the previous 'if' [-Wmisleading-indentation]

	                        return 0;
	                        ^

../fs/crypto/policy.c:634:3: note: previous statement is here

	                if (!strcmp(boot, SDHCI))

I wrongfully chose to guard the problematic return under
(!strcmp(boot, SDHCI)) by introducing brackets to the conditional,
when the proper solution should have been to simply back tab it.

As with most crypto-related code, when it fucks up, it fucks up
hard and what would appear to be a trivial modification actually
resulted in a non-booting kernel on OnePlus 8 (Pro) devices.
Ironically this did not cause any breakage on OnePlus 8T, which
happens to be the device that I owns, so I didn't even realize
there was an issue until I let the builds fly publicly.

How embarrassing.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>